moved add outfit to own component
